+++ tests/compile-test.rs
@@ -165,6 +165,7 @@ fn base_config(test_dir: &str) -> compiletest::Config {
 fn run_ui() {
     let mut config = base_config("ui");
     config.rustfix_coverage = true;
+    // use tests/clippy.toml
     let _g = VarGuard::set("CARGO_MANIFEST_DIR", fs::canonicalize("tests").unwrap());
     let _threads = VarGuard::set(
         "RUST_TEST_THREADS",
@@ -384,13 +385,17 @@ fn check_rustfix_coverage() {
     let missing_coverage_path = Path::new("target/debug/test/ui/rustfix_missing_coverage.txt");
 
     if let Ok(missing_coverage_contents) = std::fs::read_to_string(missing_coverage_path) {
-        assert!(RUSTFIX_COVERAGE_KNOWN_EXCEPTIONS.is_sorted());
+        assert!(RUSTFIX_COVERAGE_KNOWN_EXCEPTIONS.iter().is_sorted_by_key(Path::new));
 
         for rs_path in missing_coverage_contents.lines() {
-            let filename = rs_path.strip_prefix("tests/ui/").unwrap();
+            let filename = Path::new(rs_path).strip_prefix("tests/ui/").unwrap();
             assert!(
-                RUSTFIX_COVERAGE_KNOWN_EXCEPTIONS.binary_search(&filename).is_ok(),
-                "`{}` runs `MachineApplicable` diagnostics but is missing a `run-rustfix` annotation",
+                RUSTFIX_COVERAGE_KNOWN_EXCEPTIONS
+                    .binary_search_by_key(&filename, Path::new)
+                    .is_ok(),
+                "`{}` runs `MachineApplicable` diagnostics but is missing a `run-rustfix` annotation. \
+                Please either add `// run-rustfix` at the top of file or add the file to \
+                `RUSTFIX_COVERAGE_KNOWN_EXCEPTIONS`.",
                 rs_path,
             );
         }
