+++ tests/ui/needless_late_init.fixed
@@ -1,6 +1,12 @@
 // run-rustfix
 #![feature(let_chains)]
-#![allow(unused, clippy::nonminimal_bool, clippy::let_unit_value, clippy::let_and_return)]
+#![allow(
+    unused,
+    clippy::assign_op_pattern,
+    clippy::let_and_return,
+    clippy::let_unit_value,
+    clippy::nonminimal_bool
+)]
 
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::rc::Rc;
@@ -229,4 +235,24 @@ fn does_not_lint() {
     let x;
     let y = SignificantDrop;
     x = SignificantDrop;
+}
+
+mod fixable {
+    #![allow(dead_code)]
+
+    fn main() {
+        
+        let a = "zero";
+
+        
+        
+        let b = 1;
+        let c = 2;
+
+        
+        let d: usize = 1;
+
+        
+        let e = format!("{}", d);
+    }
 }
