+++ tests/ui/needless_late_init.stderr
@@ -1,5 +1,5 @@
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:16:5
+  --> $DIR/needless_late_init.rs:22:5
    |
 LL |     let a;
    |     ^^^^^^
@@ -21,7 +21,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:25:5
+  --> $DIR/needless_late_init.rs:31:5
    |
 LL |     let b;
    |     ^^^^^^
@@ -42,7 +42,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:32:5
+  --> $DIR/needless_late_init.rs:38:5
    |
 LL |     let d;
    |     ^^^^^^
@@ -63,7 +63,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:40:5
+  --> $DIR/needless_late_init.rs:46:5
    |
 LL |     let e;
    |     ^^^^^^
@@ -84,7 +84,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:47:5
+  --> $DIR/needless_late_init.rs:53:5
    |
 LL |     let f;
    |     ^^^^^^
@@ -100,7 +100,7 @@ LL +         1 => "three",
    | 
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:53:5
+  --> $DIR/needless_late_init.rs:59:5
    |
 LL |     let g: usize;
    |     ^^^^^^^^^^^^^
@@ -120,7 +120,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:61:5
+  --> $DIR/needless_late_init.rs:67:5
    |
 LL |     let x;
    |     ^^^^^^ created here
@@ -134,7 +134,7 @@ LL |     let x = 1;
    |     ~~~~~
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:65:5
+  --> $DIR/needless_late_init.rs:71:5
    |
 LL |     let x;
    |     ^^^^^^ created here
@@ -148,7 +148,7 @@ LL |     let x = SignificantDrop;
    |     ~~~~~
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:69:5
+  --> $DIR/needless_late_init.rs:75:5
    |
 LL |     let x;
    |     ^^^^^^ created here
@@ -162,7 +162,7 @@ LL |     let x = SignificantDrop;
    |     ~~~~~
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:88:5
+  --> $DIR/needless_late_init.rs:94:5
    |
 LL |     let a;
    |     ^^^^^^
@@ -183,7 +183,7 @@ LL |     };
    |      +
 
 error: unneeded late initialization
-  --> $DIR/needless_late_init.rs:105:5
+  --> $DIR/needless_late_init.rs:111:5
    |
 LL |     let a;
    |     ^^^^^^
@@ -203,5 +203,72 @@ help: add a semicolon after the `match` expression
 LL |     };
    |      +
 
-error: aborting due to 11 previous errors
+error: unneeded late initialization
+  --> $DIR/needless_late_init.rs:244:9
+   |
+LL |         let a;
+   |         ^^^^^^ created here
+LL |         a = "zero";
+   |         ^^^^^^^^^^ initialised here
+   |
+help: declare `a` here
+   |
+LL |         let a = "zero";
+   |         ~~~~~
+
+error: unneeded late initialization
+  --> $DIR/needless_late_init.rs:247:9
+   |
+LL |         let b;
+   |         ^^^^^^ created here
+LL |         let c;
+LL |         b = 1;
+   |         ^^^^^ initialised here
+   |
+help: declare `b` here
+   |
+LL |         let b = 1;
+   |         ~~~~~
+
+error: unneeded late initialization
+  --> $DIR/needless_late_init.rs:248:9
+   |
+LL |         let c;
+   |         ^^^^^^ created here
+LL |         b = 1;
+LL |         c = 2;
+   |         ^^^^^ initialised here
+   |
+help: declare `c` here
+   |
+LL |         let c = 2;
+   |         ~~~~~
+
+error: unneeded late initialization
+  --> $DIR/needless_late_init.rs:252:9
+   |
+LL |         let d: usize;
+   |         ^^^^^^^^^^^^^ created here
+LL |         d = 1;
+   |         ^^^^^ initialised here
+   |
+help: declare `d` here
+   |
+LL |         let d: usize = 1;
+   |         ~~~~~~~~~~~~
+
+error: unneeded late initialization
+  --> $DIR/needless_late_init.rs:255:9
+   |
+LL |         let e;
+   |         ^^^^^^ created here
+LL |         e = format!("{}", d);
+   |         ^^^^^^^^^^^^^^^^^^^^ initialised here
+   |
+help: declare `e` here
+   |
+LL |         let e = format!("{}", d);
+   |         ~~~~~
+
+error: aborting due to 16 previous errors
