--- tests/ui/needless_late_init_fixable.rs
@@ -1,19 +1,0 @@
-// run-rustfix
-
-#![allow(unused, clippy::assign_op_pattern)]
-
-fn main() {
-    let a;
-    a = "zero";
-
-    let b;
-    let c;
-    b = 1;
-    c = 2;
-
-    let d: usize;
-    d = 1;
-
-    let e;
-    e = format!("{}", d);
-}
