--- tests/ui/unnecessary_cast_fixable.fixed
@@ -1,50 +1,0 @@
-// run-rustfix
-
-#![warn(clippy::unnecessary_cast)]
-#![allow(
-    clippy::no_effect,
-    clippy::unnecessary_operation,
-    clippy::nonstandard_macro_braces,
-    clippy::borrow_as_ptr
-)]
-
-fn main() {
-    // casting integer literal to float is unnecessary
-    100_f32;
-    100_f64;
-    100_f64;
-    let _ = -100_f32;
-    let _ = -100_f64;
-    let _ = -100_f64;
-    100_f32;
-    100_f64;
-    // Should not trigger
-    #[rustfmt::skip]
-    let v = vec!(1);
-    &v as &[i32];
-    0x10 as f32;
-    0o10 as f32;
-    0b10 as f32;
-    0x11 as f64;
-    0o11 as f64;
-    0b11 as f64;
-
-    1_u32;
-    0x10_i32;
-    0b10_usize;
-    0o73_u16;
-    1_000_000_000_u32;
-
-    1.0_f64;
-    0.5_f32;
-
-    1.0 as u16;
-
-    let _ = -1_i32;
-    let _ = -1.0_f32;
-
-    let _ = 1 as I32Alias;
-    let _ = &1 as &I32Alias;
-}
-
-type I32Alias = i32;
